requirements.txt
selenium
requests
gym
numpy
pyautogui
Pillow
stable_baselines
tensorflow==1.8.0
stable_baselines[mpi]
tensorflow-gpu==1.8.0